Reorganize custom widgets into new module
Move preset widgets there too
Remove old sound-list widgets
Rename PresetObject to Preset
