+++ blanket/meson.build
@@ -22,16 +22,19 @@ configure_file(
   install_dir: get_option('bindir')
 )
 
+# Python sources
 sources = [
   '__init__.py',
   'main.py',
   'window.py',
   'preferences.py',
-  'presets.py',
+  'preset.py',
   'sound.py',
-  'widgets.py',
   'settings.py',
   'mpris.py'
 ]
 
 install_data(sources, install_dir: moduledir)
+
+# Install custom widgets
+install_subdir('widgets', install_dir: moduledir)
