+++ blanket/widgets/__init__.py
@@ -1,0 +1,8 @@
+# Copyright 2022 Rafael Mardojai CM
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+from blanket.widgets.playpause import PlayPauseButton # noqa
+from blanket.widgets.preset_chooser import PresetChooser # noqa
+from blanket.widgets.preset_dialog import PresetDialog # noqa
+from blanket.widgets.sound_item import SoundItem # noqa
+from blanket.widgets.volume_row import VolumeRow  # noqa
