+++ data/resources/blanket.gresource.xml
@@ -9,7 +9,6 @@
     <file>preset-row.ui</file>
     <file>about.ui</file>
     <file alias="gtk/help-overlay.ui">shortcuts.ui</file>
-    <file>sound-row.ui</file>
     <file>volume-row.ui</file>
     <file>grid-item.ui</file>
