--- data/resources/sound-row.ui
@@ -1,67 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- Generated with glade 3.22.0 -->
-<interface>
-  <template class="SoundRow" parent="GtkListBoxRow">
-    <property name="selectable">False</property>
-    <style>
-      <class name="soundrow"/>
-    </style>
-    <child>
-      <object class="GtkBox">
-        <property name="spacing">24</property>
-        <child>
-          <object class="GtkImage" id="icon">
-            <property name="pixel-size">64</property>
-          </object>
-        </child>
-        <child>
-          <object class="GtkBox">
-            <property name="hexpand">True</property>
-            <property name="vexpand">True</property>
-            <property name="valign">center</property>
-            <property name="orientation">vertical</property>
-            <child>
-              <object class="GtkBox">
-                <property name="margin_top">10</property>
-                <property name="spacing">6</property>
-                <child>
-                  <object class="GtkLabel" id="title">
-                    <property name="ellipsize">end</property>
-                    <property name="xalign">0</property>
-                  </object>
-                </child>
-              </object>
-            </child>
-            <child>
-              <object class="GtkScale" id="volume">
-                <property name="hexpand">True</property>
-                <property name="adjustment">
-                  <object class="GtkAdjustment" id="volume_adjustment">
-                    <property name="lower">0</property>
-                    <property name="upper">1</property>
-                    <property name="step_increment">0.01</property>
-                    <property name="page_increment">0.01</property>
-                  </object>
-                </property>
-                <property name="draw_value">False</property>
-              </object>
-            </child>
-          </object>
-        </child>
-        <child>
-          <object class="GtkButton" id="remove_btn">
-            <style>
-              <class name="flat"/>
-              <class name="circular"/>
-            </style>
-            <binding name="visible">
-              <lookup name="custom">SoundRow</lookup>
-            </binding>
-            <property name="icon-name">com.rafaelmardojai.Blanket-remove-symbolic</property>
-            <property name="valign">center</property>
-          </object>
-        </child>
-      </object>
-    </child>
-  </template>
-</interface>
